Publish HTTP endpoint metadata through compiler lifecycle
