+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/endpointyaml/generator/EndpointYamlGenerator.java
@@ -18,9 +18,6 @@
 
 package io.ballerina.stdlib.http.compiler.endpointyaml.generator;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator;
 import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.compiler.syntax.tree.NodeList;
 import io.ballerina.compiler.syntax.tree.ServiceDeclarationNode;
@@ -31,15 +28,11 @@ import io.ballerina.tools.diagnostics.DiagnosticSeverity;
 import io.swagger.v3.oas.models.servers.Server;
 import io.swagger.v3.oas.models.servers.ServerVariables;
 
-import java.io.IOException;
 import java.io.PrintStream;
-import java.io.Writer;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 
 /*
- * Computes endpoint details of an HTTP service and writes the consolidated endpoints.yaml
+ * Computes endpoint details of an HTTP service
  */
 public class EndpointYamlGenerator {
     private final ServiceDeclarationNode node;
@@ -47,9 +40,7 @@ public class EndpointYamlGenerator {
 
     private static final PrintStream outStream = System.out;
 
-    private static final String ARTIFACT = "artifact";
     private static final String REST = "REST";
-    private static final String ENDPOINTS_FILE_NAME = "endpoints.yaml";
     private static final String PORT = "port";
 
     private int portVal = 0;
@@ -79,22 +70,16 @@ public class EndpointYamlGenerator {
         var portVar = vars != null ? vars.get(PORT) : null;
         String defaultPort = portVar != null ? portVar.getDefault() : null;
 
-       if (defaultPort != null && !defaultPort.isEmpty()) {
-           try {
-               this.portVal = Integer.parseInt(defaultPort);
-           } catch (NumberFormatException ex) {
-               outStream.println("Assign a integer value for port.");
-           }
-       } else {
+        if (defaultPort != null && !defaultPort.isEmpty()) {
+            try {
+                this.portVal = Integer.parseInt(defaultPort);
+            } catch (NumberFormatException ex) {
+                outStream.println("Assign a integer value for port.");
+            }
+        } else {
             diagnostics.add(missingPortConfigDiagnostic());
-       }
+        }
         return new Endpoint(basePath, this.portVal, basePath, REST, this.schemaFileName);
-    }
-
-    public static void writeEndpointsYaml(Path outPath, List<Endpoint> endpoints) throws IOException {
-        Files.createDirectories(outPath.resolve(ARTIFACT));
-        Path path = outPath.resolve(ARTIFACT).resolve(ENDPOINTS_FILE_NAME);
-        writeYaml(path, new EndpointsWrapper(endpoints));
     }
 
     private String getBasePath() {
@@ -104,20 +89,6 @@ public class EndpointYamlGenerator {
             serviceBasePath.append(identifierNode.toString().replace("\"", "").trim());
         }
         return serviceBasePath.toString();
-    }
-
-    private static void writeYaml(Path path, EndpointsWrapper wrapper) throws IOException {
-        YAMLFactory yamlFactory = YAMLFactory.builder()
-                .disable(YAMLGenerator.Feature.WRITE_DOC_START_MARKER)
-                .build();
-        ObjectMapper mapper = new ObjectMapper(yamlFactory);
-        mapper.findAndRegisterModules();
-
-        try (Writer writer = Files.newBufferedWriter(path)) {
-            mapper.writeValue(writer, wrapper);
-        } catch (IOException e) {
-            throw new IOException("Failed to write endpoints yaml to " + path, e);
-        }
     }
 
     private Diagnostic missingPortConfigDiagnostic() {
