+++ compiler-plugin/src/main/java/io/ballerina/stdlib/http/compiler/endpointyaml/generator/ServiceArtifactsLifecycleListener.java
@@ -24,8 +24,8 @@ import io.ballerina.projects.plugins.CompilerLifecycleListener;
 import java.util.Map;
 
 /*
- * Registers the {@code EndpointsYamlWriterTask} to run once the whole compilation's code generation
- * has completed, so it can write every service's collected endpoint details into a single endpoints.yaml.
+ * Registers the {@code EndpointMetadataTask} to run once the whole compilation's code generation
+ * has completed, so it can publish every service's collected endpoint metadata to Ballerina lang.
  */
 public class ServiceArtifactsLifecycleListener extends CompilerLifecycleListener {
     private final Map<String, Object> ctxData;
@@ -36,6 +36,6 @@ public class ServiceArtifactsLifecycleListener extends CompilerLifecycleListener
 
     @Override
     public void init(CompilerLifecycleContext context) {
-        context.addCodeGenerationCompletedTask(new EndpointsYamlWriterTask(ctxData));
+        context.addCodeGenerationCompletedTask(new EndpointMetadataTask(ctxData));
     }
 }
